test-requirements.txt
pyyaml
pytest
bcrypt
juju
black
flake8
